Move extraction of precursor from transaction into proper module
