+++ rust/main/submitter/src/chain_tx_adapter/chains/sealevel/transaction/precursor.rs
@@ -1,0 +1,15 @@
+use crate::chain_tx_adapter::SealevelTxPrecursor;
+use crate::transaction::{Transaction, VmSpecificTxData};
+
+pub trait Precursor {
+    fn precursor(&self) -> &SealevelTxPrecursor;
+}
+
+impl Precursor for Transaction {
+    fn precursor(&self) -> &SealevelTxPrecursor {
+        match self.vm_specific_data() {
+            VmSpecificTxData::Svm(precursor) => precursor,
+            _ => panic!(),
+        }
+    }
+}
